requirements.txt
Django==4.0.6
django-admin-sortable2==2.1
django_database_url==1.0.3
environs==9.5.0
Pillow==9.2.0